Changes to total_waiting_time in intersection:get_wait () - to calculate average waiting time by each car and minimizing that. But doesn't work displayPerformance uncommented in Learn&Comply Mode

+++ src/intersection.h
@@ -49,10 +49,10 @@ class intersection
 
   int number_of_actions_per_state;
 
-  int curr_wait;
-  int prev_wait;
+  float curr_wait;
+  float prev_wait;
 
-  int total_waiting_cars;
+  float total_waiting_cars;
   float reward;
   int action;
   int best_action;
